modifyed task_9

+++ week_03/20210113_9.c
@@ -12,15 +12,15 @@
 
 int main(){
 
-    float valueOne = 4.9876543; //chaged the variable type from double to float.
-    double valueTwo = 7.123456789012345678890;
-    long double valueThree = 18398458438583853.28;
-    long double valueFour = 18398458438583853.39875937284928422;
+    float valueOne = 4.9876543f; //chaged the variable type from double to float.
+    float valueTwo = 7.123456789012345678890F;
+    long double valueThree = 18398458438583853.28L;
+    long double valueFour = 18398458438583853.39875937284928422L;
 
     printf("The first value is: %.7f \n", valueOne);
     printf("The second value is: %.21f \n", valueTwo);
     printf("The third value is: %.2Lf \n", valueThree);
-    printf("The fourth value is: %Lf \n", valueFour);
+    printf("The fourth value is: %.12Lf \n", valueFour);
 
 return 0;
 }
